+++ ChangeLog.md
@@ -20,6 +20,8 @@ See docs/process.md for more on how version tagging works.
 
 3.2.0
 -----
+- emcc not accepts `-print-file-name` and reports the correct library paths in
+  `-print-search-dirs`.
 - `tools/file_packager` no longer generates (or requires) any "pre-js" code when
   running in `--embed-file` mode.  Instead the embedded files are loaded at
   static constructor time.